Add debounce counter
Based on Kay's suggestion, debounce mechanism should entail time and
number of occurances. This patch introduces two new fields, CNTx and MDCx
for tracking number of times particular link has alarmed and to set the
threshold when that propagates to the record level.

DLYx and CNTx now provide link alarm accounting regardless whether link
is enabled or not.